chore: extract logic to util function
